+++ library_filter.py
@@ -1,10 +1,3 @@
-import numpy as np
-#import pylab as py
-import glob, os
-import os.path
-import sys
-from HL2_params import *
-
 #Return a list of Kmers that cover "threshold" % of the primerss
 def kmers_threshold(primer, threshold):
 
@@ -267,7 +260,7 @@ def remove_duplicates2(library_name, run_name):
 
 			# compare last 20bp of sequences to "seen" set
 			sequence_end = sequenceLine[-20:] # make variable
-			
+
 			if sequence_end not in unique_set: # TODO: compare hamming distances
 				duplicates_removed.write(header+sequenceLine)
 				unique_set.append(sequence_end)
@@ -276,7 +269,7 @@ def remove_duplicates2(library_name, run_name):
 		print(unique_set[1:5])
 		print(lt+"\t\tdups: " + str(duplicate_count))
 		print(lt+"\t\tunique: " + str(len(unique_set)))
-		
+
 def remove_duplicates(library_name, run_name):
 		#open trimmed file & remove duplicates
 		trimmed_file_path = "results/"+run_name+'/'+library_name+"/"+library_name+".trimmed.fastq"
@@ -295,47 +288,14 @@ def remove_duplicates(library_name, run_name):
 
 			# grab last 20bp
 			sequence_end = sequenceLine[-20:] # make variable
-			
+
 			# compare to "seen" set
 			a = [repseq[-20:] for repseq in unique_set] #<- list of -20 ends.
 			b = [(repseq[-20:] == sequence_end) for repseq in unique_set] #<- T/F matching
 			next((s for s in mylist if sub in s), None)
 			c = next([repseq[-20:] == sequence_end for repseq in unique_set],None)
 			repseq = next((s for s in rep_ends_set if sequence_end in s), None) #<- rep. end matching sequence end
-			if repseq is not None and len(sequenceLine) > 
-				
+			if repseq is not None and len(sequenceLine)
+
 			# replace curruent representative sequence if not longest
-			if repseq is not None and len(sequenceLine) >
-
-
-def create_library(r1, r2, library_name, run_name):
-
-	#if the library ('results/run_name/library_name') already exists, move to next library.
-	if os.path.exists('results/'+run_name+"/"+library_name):
-		print(library_name + ' already exists... skipping!')
-
-	#Otherwise, make a library folder and fill it up with relevant library files
-	else:
-		os.makedirs('results/'+run_name+"/"+library_name)
-		print(lt+'Making ' + library_name + '...')
-		merge_reads(r1, r2, library_name, run_name)
-		collate(library_name, run_name)
-		#feature_count(FEATURES, library_name, run_name)
-		feature_trim(FEATURES, library_name, run_name)
-		remove_duplicates(library_name, run_name)
-
-if __name__ == "__main__":
-
-	# set logo type (lt) and run name variable
-	lt = "[TEASEQ] "
-	RUN_NAME = sys.argv[1]
-
-	print(lt+"Checking parameters...")
-    	validate_parameters()
-
-	print(lt+"Creating required file paths...")
-	create_paths(RUN_NAME)
-
-	print("Creating libraries...")
-	for lib in range(len(LIBS)): #create library from each pair of fastq files...
-		create_library(R1_FILES[lib], R2_FILES[lib], LIBS[lib], RUN_NAME)
+			if repseq is not None and len(sequenceLine): break
